third_party/blink/web_tests/platform/mac/external/wpt/css/css-shadow-parts/grouping-with-checked-expected.txt

This is a testharness.js-based test.
[FAIL] Styles applied to ::part(...):checked
  assert_equals: expected "rgb(0, 255, 0)" but got "rgb(255, 0, 0)"
[FAIL] Styles applied via grouped selector including matched ::part(...):checked
  assert_equals: expected "rgb(0, 255, 0)" but got "rgb(0, 0, 0)"
[FAIL] Styles applied via grouped selector including unmatched ::part(...):checked
  assert_equals: expected "monospace" but got "Times"
Harness: the test ran to completion.
